requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.13.1
bcrypt==4.1.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
docutils==0.20.1
idna==3.6
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
paramiko==3.3.1
pycparser==2.21
Pygments==2.17.2
PyNaCl==1.5.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.2
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
